Вынес функции аутентификации из routes в отдельный файл
