+++ src/node/api/__init__.py
@@ -1,2 +1,2 @@
 
-from . import routes, routes_debug
+from . import routes, routes_debug, routes_auth
